image-background-remover/requirements_image_processing.txt
# Requirements for black background removal script
Pillow>=9.0.0
numpy>=1.21.0
